# api/openapi/swagger.json
{
    "components": {
        "schemas": {
...
            "github_com_donaldgifford_server-price-tracker_pkg_types.ComponentType": {
                "description": "Extracted data",
                "enum": [
                    "ram",
                    "drive",
                    "server",
                    "cpu",
                    "nic",
                    "other"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "ComponentRAM",
                    "ComponentDrive",
                    "ComponentServer",
                    "ComponentCPU",
                    "ComponentNIC",
                    "ComponentOther"
                ]
            },
            "github_com_donaldgifford_server-price-tracker_pkg_types.Condition": {
                "enum": [
                    "new",
                    "like_new",
                    "used_working",
                    "for_parts",
                    "unknown"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "ConditionNew",
                    "ConditionLikeNew",
                    "ConditionUsedWorking",
                    "ConditionForParts",
                    "ConditionUnknown"
                ]
            },
            "github_com_donaldgifford_server-price-tracker_pkg_types.Listing": {
                "properties": {
                    "attributes": {
                        "additionalProperties": {},
                        "type": "object"
                    },
                    "component_type": {
                        "$ref": "#/components/schemas/github_com_donaldgifford_server-price-tracker_pkg_types.ComponentType"
                    },
                    "condition_norm": {
                        "$ref": "#/components/schemas/github_com_donaldgifford_server-price-tracker_pkg_types.Condition"
                    },
                    "condition_raw": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "ebay_item_id": {
                        "type": "string"
                    },
                    "extraction_confidence": {
                        "type": "number"
                    },
                    "first_seen_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "image_url": {
                        "type": "string"
                    },
                    "item_url": {
                        "type": "string"
                    },
                    "listed_at": {
                        "description": "Timestamps",
                        "type": "string"
                    },
                    "listing_type": {
                        "$ref": "#/components/schemas/github_com_donaldgifford_server-price-tracker_pkg_types.ListingType"
                    },
                    "price": {
                        "description": "Pricing",
                        "type": "number"
                    },
                    "product_key": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "score": {
                        "description": "Scoring",
                        "type": "integer"
                    },
                    "score_breakdown": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "seller_feedback_pct": {
                        "type": "number"
                    },
                    "seller_feedback_score": {
                        "type": "integer"
                    },
                    "seller_name": {
                        "description": "Seller",
                        "type": "string"
                    },
                    "seller_top_rated": {
                        "type": "boolean"
                    },
                    "shipping_cost": {
                        "type": "number"
                    },
                    "sold_at": {
                        "type": "string"
                    },
                    "sold_price": {
                        "type": "number"
                    },
                    "title": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "github_com_donaldgifford_server-price-tracker_pkg_types.ListingType": {
                "enum": [
                    "auction",
                    "buy_it_now",
                    "best_offer"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "ListingAuction",
                    "ListingBuyItNow",
                    "ListingBestOffer"
                ]
            },
...
            "internal_api_handlers.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "something went wrong",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "internal_api_handlers.listingsResponse": {
                "properties": {
                    "limit": {
                        "type": "integer"
                    },
                    "listings": {
                        "items": {
                            "$ref": "#/components/schemas/github_com_donaldgifford_server-price-tracker_pkg_types.Listing"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "offset": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "contact": {
            "name": "Donald Gifford",
            "url": "https://github.com/donaldgifford/server-price-tracker"
        },
        "description": "API for monitoring eBay server hardware listings, extracting structured attributes via LLM, scoring deals, and sending alerts.",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "title": "Server Price Tracker API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/api/v1/listings": {
            "get": {
                "description": "Returns listings with optional filters for component type, score range, and pagination.",
                "parameters": [
                    {
                        "description": "Filter by component type",
                        "in": "query",
                        "name": "component_type",
                        "schema": {
                            "enum": [
                                "ram",
                                "drive",
                                "server",
                                "cpu",
                                "nic",
                                "other"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by product key",
                        "in": "query",
                        "name": "product_key",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Minimum composite score",
                        "in": "query",
                        "name": "min_score",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Maximum composite score",
                        "in": "query",
                        "name": "max_score",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Number of results (default 50)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Pagination offset",
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Sort field",
                        "in": "query",
                        "name": "order_by",
                        "schema": {
                            "enum": [
                                "score",
                                "price",
                                "first_seen_at"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_api_handlers.listingsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_api_handlers.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_api_handlers.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List listings",
                "tags": [
                    "listings"
                ]
            }
        },
        "/api/v1/listings/{id}": {
            "get": {
                "description": "Returns a single listing by its UUID.",
                "parameters": [
                    {
                        "description": "Listing UUID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_donaldgifford_server-price-tracker_pkg_types.Listing"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_api_handlers.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_api_handlers.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Get a listing by ID",
                "tags": [
                    "listings"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Local development server",
            "url": "http://localhost:8080"
        }
    ]
}